Move inline function to .cpp to prevent compilation error
