+++ global_defines.h
@@ -31,15 +31,7 @@ struct device_address {
     }
 };
 
-void printDeviceAddress(device_address *address) {
-    for (uint8_t i = 0; i < sizeof(device_address); i++) {
-        if (i == sizeof(device_address) - 1) {
-            Serial.print(address->bytes[i]);
-        } else {
-            Serial.print(address->bytes[i]);
-            Serial.print(", ");
-        }
-    }
-}
+
+extern void printDeviceAddress(device_address *address);
 
 #endif //ARDUINO_MQTTSN_CLIENT_GLOBAL_DEFINES_H
